fix: address CodeRabbit round-2 findings on approval pagination and auth/ws races
